adiciona metodo PUT e endpoint /help

+++ aula-06/controller/filme/controller_filme.js
@@ -2,7 +2,7 @@
  * Objetivo: Arquivo responsável pela validação, tratamento e manipulação de dados para o CRUD de Filmes
  * Data: 17/04/2026
  * Autor: Julio Augusto
- * Versão: 1.0
+ * Versão: 1.0.4.26
  * *******************************************************************************************************/
 
 // Import do arquivo de padronização de mensagens
@@ -25,22 +25,20 @@ const inserirNovoFilme = async (filme, contentType) => {
         let validar = await validarDados(filme)
 
         // Se a função 'validarDados' retornar um JSON de erro, iremos retornar o erro ao app
-        if(validar){
+        if(validar)
             return validar // Status code 400
-
-        }else { 
-            // tenta inserir no banco
-            let result = await filmeDAO.insertFilme(filme)
+        
+        // tenta inserir no banco
+        let result = await filmeDAO.insertFilme(filme)
                 
-            if(result){ 
-                message.DEFAULT_MESSAGE.status = message.SUCESS_CREATED_ITEM.status
-                message.DEFAULT_MESSAGE.status_code = message.SUCESS_CREATED_ITEM.status_code
-                message.DEFAULT_MESSAGE.response = message.SUCESS_CREATED_ITEM.message
-                return message.DEFAULT_MESSAGE // Status code 201
-
-            } 
-            return message.ERROR_INTERNAL_SERVER_MODEL // Status code 500
+        if(result){ 
+            message.DEFAULT_MESSAGE.status = message.SUCESS_CREATED_ITEM.status
+            message.DEFAULT_MESSAGE.status_code = message.SUCESS_CREATED_ITEM.status_code
+            message.DEFAULT_MESSAGE.response = message.SUCESS_CREATED_ITEM.message
+            return message.DEFAULT_MESSAGE // Status code 201
         }
+
+        return message.ERROR_INTERNAL_SERVER_MODEL // Status code 500
 
     } catch (error) {
         return message.ERROR_INTERNAL_SERVER_CONTROLLER // Status code 500  
@@ -48,8 +46,45 @@ const inserirNovoFilme = async (filme, contentType) => {
 }
 
 // Função para atualizar um filme
-const atualizarFilme = async (filme) => {
-    
+const atualizarFilme = async (filme, id, contentType) => {
+    let message = JSON.parse(JSON.stringify(config_message))
+
+    try {
+        // Valida se o formato de dados é JSON
+        if(String(contentType).toLowerCase() != 'application/json')
+            return message.ERROR_CONTENT_TYPE // Status code 415
+
+        let resultBuscarID = await buscarFilme(id)
+
+        // Valida se é possivel encontrar o filme
+            // Se o status for true, o filme foi encontrado
+            // Se for false, o filme não foi encontrado ou houve um erro de processamento
+        if(!resultBuscarID.status)
+            return resultBuscarID // Status code 400 ou 404 ou 500
+
+        // valida dados obrigatórios
+        let validar = await validarDados(filme)
+        if(validar)
+            return validar
+        
+        // adiciona atributo 'id' do filme no JSON para ser enviado no DAO
+        filme.id = id
+        let result = await filmeDAO.updateFilme(filme) // Tenta update no banco
+        
+        // retorno sucesso
+        if(result){
+            message.DEFAULT_MESSAGE.status = message.SUCESS_UPDATE_ITEM.status
+            message.DEFAULT_MESSAGE.status_code = message.SUCESS_UPDATE_ITEM.status_code
+            message.DEFAULT_MESSAGE.message = message.SUCESS_UPDATE_ITEM.message
+            return message.DEFAULT_MESSAGE // Status code 200
+        }
+
+        // Erro na model
+        return message.ERROR_INTERNAL_SERVER_MODEL // Status code 500
+
+    } catch (error) {
+        return message.ERROR_INTERNAL_SERVER_CONTROLLER // Status code 500  
+    }
 }
 
 // Função para retornar Todos filmes
@@ -86,7 +121,7 @@ const buscarFilme = async (id) => {
     try {
         // tratamentos dados incorretos
         if (id.trim() == '' || id == null || id == undefined || id < 1 || isNaN(id)){
-            message.DEFAULT_MESSAGE.field = '[ID] INVÁLIDO'
+            message.ERROR_BAD_REQUEST.field = '[ID] INVÁLIDO'
             return message.ERROR_BAD_REQUEST // status_code 400
         }
         // executa a função para retornar um filme pelo id
@@ -119,18 +154,26 @@ const excluirFilme = async (id) => {
     try {
         // tratamentos dados incorretos
         if (id.trim() == '' || id == null || id == undefined || id < 1 || isNaN(id)){
-            message.DEFAULT_MESSAGE.field = '[ID] INVÁLIDO'
+            message.ERROR_BAD_REQUEST.field = '[ID] INVÁLIDO'
             return message.ERROR_BAD_REQUEST // status_code 400
         }
+
+        // Valida se é possivel encontrar o filme
+        let resultBuscarID = await buscarFilme(id)
+        if(!resultBuscarID.status)
+            return message.ERROR_NOT_FOUND
 
         // executa a função que deleta um filme pelo id no banco de dados
         let result = await filmeDAO.deleteFilme(id)
 
         if(result){
-            return message.SUCESS_RESPONSE // status_code 200
-        }else{
-            return message.ERROR_INTERNAL_SERVER_MODEL // status_code 500
+            message.DEFAULT_MESSAGE.status = message.SUCESS_DELETE_ITEM.status
+            message.DEFAULT_MESSAGE.status_code = message.SUCESS_DELETE_ITEM.status_code
+            message.DEFAULT_MESSAGE.message = message.SUCESS_DELETE_ITEM.message
+            return message.DEFAULT_MESSAGE // status_code 200
         }
+        
+        return message.ERROR_INTERNAL_SERVER_MODEL // status_code 500
         
     } catch (error) {
         return message.ERROR_INTERNAL_SERVER_CONTROLLER // status_code 500
@@ -166,8 +209,7 @@ const validarDados = async (filme) => {
         message.ERROR_BAD_REQUEST.field = '[CAPA] INVÁLIDO'
         return message.ERROR_BAD_REQUEST
     }else{
-        return false
-        
+        return false  
     }
 }
